Split updateSheets and updateValues functions and merged balanceSheet and balanceSheetList
updateAssetsSheet and updateLiabilitiesSheet are now used for adding new fields to the assets and liabilities templates

updateAssetValues and updateLiabilityValues are used for updating the values

the reasoning here was that values and templatese would need to be seperate for sending to the back end and for performing additional calculations such as totals, so this seemed like an obvious change. the added benifit is that we now also have given the user the ability to add new fields

balance sheet list is now rolled into balance sheet becasue balance sheet list only ever exists inside of balance sheets and thus doesn't need to be seperate. having both separate made more sense