Retrieval and Updating are Split
Retrieval of values from backend and updating the store in the front end needed to be split because updating the store would load new values the user couldn't forsee and could potentially lead to a session being destroyed by auto loading.

Another issue was that when new values were brought in the totalAssets, totalLiabilities and net worth values would need to be updated. If the front end had changes the backend didn't and vice versa then these values would become invalid and the only way to rectify them would be to recalculate them entirely. Recalculation is possible but the previous iteration updates would be meaningless because retrieval calls are made on every value update, thus recalculation would occur on every value update. Recalculation would also require the server's value state and the store's value state to be sent to the total reducers, which seems like a bad approach.

These two factors: unexpected data being pushed on the user and greatly increased complexity in total reducers led me to settle on a new solution.

The new paradigm will be that data is retrieved when the app component is constructed. After that point the user will be able to push changes by modifying the front end but will only see backend changes by calling the load functions.

This updated paradigm still meets requirements by maintaining an ultimate source of truth on the backend that can be modified by the front end. If the front end store ever surpasses the backend it will be only in temporary holdings which are lost on refresh unless the user commits them to the backend.

We also now have a technically easy and user intuitive upgrade path for many users simultaneously working on a single document. Users modify their local version and send it to the backend where it can be stored as a temp document. Other users can pull temp documents and preview the changes before merging with their locals instead of having the front end auto merge and potentially cause undesired behaviour.