docId added and valueType added
valueType replaces isAsset and isLiability to make sure backend is 3NF. docId is no longer hard coded allowing for sheets to be changed from front end. unfortunately there is still an error where sheet values won't visibly update because their value is initialized to zero and thus they are uncontrolled. the solution to this is to set up 0'd values in assetvalues and liability values when calls are made to update or retrieve the sheets. this will be done in a future commit